+++ src/NL/nlFile.cpp
@@ -83,7 +83,7 @@ void nlLoadEntireFileAsyncCallback(nlFile* file, void* pBuffer, unsigned int uSi
 /**
  * Offset/Address/Size: 0x160 | 0x801CEB90 | size: 0xDC
  */
-void* nlLoadEntireFile(const char* filename, unsigned long* outSize, unsigned int alignment, eAllocType type)
+void* nlLoadEntireFile(const char* filename, u32* outSize, unsigned int alignment, eAllocType type)
 {
     unsigned int datasize;
     nlFile* file;
